requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.6
jinja2==3.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
qrcode==7.4.2
pillow==10.1.0
numpy==1.24.3
opencv-python-headless==4.8.0.76
pyzbar==0.1.9
python-dotenv==1.0.0 